Bugfix: No more nameless calculation setups
At some point during the init phase of the calculation_setup tab and
tables, the cs selected signal is emited before the CSListWidget has
loaded the calulation setups and therefore writes a cs with an empty
string as name. Because I couldn't figure out where this happend I
added a check in the controller.