Bugfix: Remove nasty signal loop that overwrites calculation setup methods
QTableWidget.cellChanged signal is emited when a cell is changed by the
user or programmatically. Therefore whenever sync was called on the
activity table, calculation_setup_changed was emited and the method
table was overwritten before it was synced, because it is triggerd by the
same signal as the activity table. This lead to all the strange
behaviour with buggy cs rename and cs delete etc.

The fix is to disconnect the cellChanged signal during sync of the
activity table.