add location list combo to activity tab allows user to select from all location names which already exist in the database of the selected activity Or add a new location that's not in the list, simply by typing
whilst implementing - found an issue that was causing many issues in the past and in this function implementation...
The "activity" data was redundantly duplicated in objects across the interface:
	addressed in this commit: ActivityTab.activity and ActivityDataGrid.activity redundancy
	not addressed in this commit: the *possible* redundancy between the ActivityCache and ActivitiesPanel, which is less clear-cut

ActivityDataGrid.activity is no longer stored. Rather, the ActivityTab is referenced as a parent, and the ActivityTab.activity object used in all code
Such that when data is being edited, it is no longer necessary to maintain synchronisation of these two separate objects, as well as the displayed values in the widgets

Also addressed the assumption(?) in previous code, that an ActivityTab instance could be created without an activity_key. This might have been a legacy thing.
	No longer provides default values for activities
	Or checks if an activity was passed (e.g. before populate()

rename all references to 'dropdown' to 'combo' - aligned with PyQT naming convention