Update search when changing a db/activity; keep heights right
Whenever a database or activity gets modified, the search query is
executed again. Also, whenever a search query gets executed, the
splitter height gets recalculated to ensure heights remain correct.

I am not entirely sure this is the right way to do the latter, though.
There seems to be a single frame of the splitter height being at some
other height, whenever first loading a table, or when performing a
search action that filters literally everything (e.g. search for some
random string of characters). This implies there is some other piece of
code setting the splitter height initially, which is then later
overwritten by resize_splitter.

This might be something to improve and actually do the 'correct' way
some time in the future.

refs #785