Remove test that isn't needed; include fix for final important to-do
The test I removed was not functioning with the changed implementation
of the uncertainty editing. Now it creates the uncertainty wizard
instead of creating a dropdown to change the uncertainty type; the test
that checks for the dropdown no longer serves a purpose.

I also included the final imporant TODO I had left, considering there
was an issue with the read-only status not updating correctly when
changing the status of the uncertainty column's visibility.

refs #482