Expand the view to include the full width of the graph minimap
Whenever a minimap is wider than the screen, the minimap will dictate
the width of the graph viewer. This means that the minimap's viewfinder
will also be much wider, and not everything indicated in the minimap's
viewfinder can be shown in your window at the same time.

An alternative solution would be to split up the minimap and the graph
itself, however that would require reworking of a significant part of
the code.

Also, this javascript code could use a cleanup, imho.

refs #924