Feature edit ICs (#999)
* Improve editability of Impact Categories

Users can now:
- Only edit ICs when the appropriate checkbox is checked
- Edit amounts and uncertainties for a CF by double clicking the cell
- Open the uncertainty wizard by double clicking the uncertainty cell
- Drag and drop biosphere activities into an impact category
-- Default amount is set to 0; this can be edited later through double
   clicking as mentioned above.

refs #482

WIP:
Duplicating entire IC trees is not yet working properly. This could
however be moved into another PR if desired.
I also spotted a minor issue that makes cells editable by double
clicking, if the 'Hide uncertainty columns' checkbox is toggled.

Also, removing a CF after adding it

* Remove test that isn't needed; include fix for final important to-do

The test I removed was not functioning with the changed implementation
of the uncertainty editing. Now it creates the uncertainty wizard
instead of creating a dropdown to change the uncertainty type; the test
that checks for the dropdown no longer serves a purpose.

I also included the final imporant TODO I had left, considering there
was an issue with the read-only status not updating correctly when
changing the status of the uncertainty column's visibility.

refs #482

* Several additional features for modifying ICs and CFs

- Ability to delete an IC (e.g. after incorrectly duplicating one)
- Ability to delete a CF from a method (e.g. after inserting one)
-- add new icon to distinguish clearing uncertainty vs deleting a CF
- Bugfix causing all opened methods to have their uncertainty wizard
  opened when doubleclicking on one method to modify the uncertainty.
- Remove automatic search query when duplicating method (still needs
  review or some more work.)
- Fix minor bug related to editability

refs #482

* Minor wording changes

refs #482

* Provides updates to the IC code supporting the deletion and duplication of sets of the IC methods, in addition to providing further functionality supporting the users interaction with the IC TreeView. Facilitating user manipulation.

* Added functionality to the TableView providing a contextMenu item for inspecting the categories in addition to the double click.

---------

Co-authored-by: Pim Bax <joeytje50@gmail.com>