Updated routine for the SuperstructureMLCA class update_matrices method. (#1023)
Used to update the technosphere and biosphere matrices to include the values from Scenarios. Updates now involve the use of copied matrices so that the original database values are not lost and can be used on repeated iterations of scenario calculations. Expected behaviour is, therefore, that each scenario will use self defined values and the values from the database where no value is defined.